SEQNG-285 Move the facades to its own module to support custom compiler params #time 1h
